Set the TTL for local answers
The local answers are typically the blacklisted domains and hostnames. So our browsers can definitively keep the answers and for good ! So the clients won't come to interrogate the Dnsmasq server so often. 
There is drawback: When there are domains/hostnames removed from the list ( because not ads anymore, or correction, or u want to whitelist!), each client that use ur dnsmasq server will need sometimes to get the update. Or you'll have to flushdns on every client, not very convenient especially on smartphones and tablets.

That's why I set 1h, I think it's a good compromise.